Switch to Backdrop core-provided Diff Engine. Fixes #27
The Diff Engine that ships with this module is old and includes several
compatibility issues, including the one referenced in Issue #26. The
diff module requires a `BackdropDiffInline` class that the core Diff
Engine doesn't include, so I've moved that into it's own class and
removed the old DiffEngine and related classes.